Release 2024-04-11

+++ packages/graphics/mesa/package.mk
@@ -11,16 +11,22 @@ PKG_PATCH_DIRS+=" ${DEVICE}"
 
 case ${DEVICE} in
   RK3588*)
-	PKG_VERSION="120202c675749c5ef81ae4c8cdc30019b4de08f4"
-	PKG_SITE="https://gitlab.com/panfork/mesa"
+	PKG_VERSION="832c3c7117e4366e415ded92a6f07ec203fd9233"
+	PKG_SITE="https://github.com/ROCKNIX/mesa-panfork"
 	PKG_URL="${PKG_SITE}.git"
-	PKG_GIT_CLONE_BRANCH="csf"
   ;;
-  RK3*) #Using upstream dev for panfrost
-	PKG_VERSION="db29984c254f60f5daeec0ea4e6048b6ee7902f8"
-	PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
-	PKG_URL="${PKG_SITE}.git"
-	PKG_PATCH_DIRS+=" panfrost"
+  RK3*|S922X)
+    if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+      PKG_VERSION="24.0.4"
+	    PKG_SITE="http://www.mesa3d.org/"
+	    PKG_URL="https://gitlab.freedesktop.org/mesa/mesa/-/archive/mesa-${PKG_VERSION}/mesa-mesa-${PKG_VERSION}.tar.gz"
+    else
+      #Using upstream dev for panfrost
+	    PKG_VERSION="db29984c254f60f5daeec0ea4e6048b6ee7902f8"
+	    PKG_SITE="https://gitlab.freedesktop.org/mesa/mesa"
+	    PKG_URL="${PKG_SITE}.git"
+	    PKG_PATCH_DIRS+=" panfrost"
+    fi
   ;;
   *)
 	PKG_VERSION="24.0.4"
@@ -34,7 +40,7 @@ get_graphicdrivers
 PKG_MESON_OPTS_TARGET="-Dgallium-drivers=${GALLIUM_DRIVERS// /,} \
                        -Dgallium-extra-hud=false \
                        -Dgallium-omx=disabled \
-                       -Dgallium-nine=false \
+                       -Dgallium-nine=true \
                        -Dgallium-opencl=disabled \
                        -Dgallium-xa=disabled \
                        -Dshader-cache=enabled \
@@ -106,9 +112,7 @@ else
 fi
 
 post_makeinstall_target() {
-  case ${DEVICE} in
-    S922X)
-      rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
-    ;;
-  esac
+  if [ "${DEVICE}" = "S922X" -a "${USE_MALI}" != "no" ]; then
+    rm -f ${INSTALL}/usr/lib/libvulkan_panfrost.so ${INSTALL}/usr/share/vulkan/icd.d/panfrost_icd.aarch64.json
+  fi
 }
